backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.40.0
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
celery[redis]==5.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
